Wil/Henrique/Coelho: Refatoração dos renders! Estamos tentando abusar dos recursos do OpenGL
git-svn-id: http://xp-dev.com/svn/horus_eye/trunk@581 863e7ce6-8861-447e-8dcb-03987b876380

+++ src/framework/frame.h
@@ -20,30 +20,30 @@ class Frame {
   public:
     // Construtores e destrutores
     Frame() : left_(0), top_(0), right_(0), bottom_(0) { }
-    Frame(int left, int top, int right, int bottom) :
+    Frame(float left, float top, float right, float bottom) :
         left_(left), top_(top), right_(right), bottom_(bottom) { }
     ~Frame() { }
 
     // Utilidades inline
-    int width() const { return right_ - left_; }
-    int height() const { return bottom_ - top_; }
+    float width() const { return right_ - left_; }
+    float height() const { return bottom_ - top_; }
 
     // Accessors e mutators
-    int left() const { return left_; }
-    void set_left(int left) { left_ = left; }
-    int top() const { return top_; }
-    void set_top(int top) { top_ = top; }
-    int right() const { return right_; }
-    void set_right(int right) { right_ = right; }
-    int bottom() const { return bottom_; }
-    void set_bottom(int bottom) {bottom_ = bottom; }
+    float left() const { return left_; }
+    void set_left(float left) { left_ = left; }
+    float top() const { return top_; }
+    void set_top(float top) { top_ = top; }
+    float right() const { return right_; }
+    void set_right(float right) { right_ = right; }
+    float bottom() const { return bottom_; }
+    void set_bottom(float bottom) {bottom_ = bottom; }
 
     // Devolve true se houver colisao entre
     // 'frame' e este retangulo.
     bool Collides(const Frame& frame) const;
 
   private:
-    int left_, top_, right_, bottom_;
+    float left_, top_, right_, bottom_;
 };
 
 }  // namespace framework
